Correctly kill child processes on exit
I added a wrapper, watch-markdown-gui, for when we get invoked from the
Gnome GUI. This was intended to redirect stdout & err, on the theory
that these are not available to scripts invoked from the GUI, and this
was causing processes to wedge and contributing to the leftover child
processes after we exit.

However, adding this wrapper didn't fix the leftover processes. I intend
to try removing the wrapper again after this commit.

What did work was tweaking watch-markdown's trap to cleanup children on
exit.

Removed a needless subshell ( ) around the inotifywait while...done
loop. This resulted in one fewer child processes.

Also, we explicitly handle a lot of error conditions around the given
.md file on startup. I perhaps ought to see what happens if I omit
those, and just pass the given file to pandoc.

Fixed the filename check within the while...done loop, to use the
directory of the filename which recieved an event. Otherwise, when
invoked from the GUI, 'realpath' was assuming filename 'x' was actually
'$(pwd)/x', where pwd=~, and hence was never comparing equal to the
expected filename.