Install executables to ~/.local/bin
This is a more standard location than ~/bin, it already exists and is on
the PATH for everyone.

Add an 'uninstall' script, which I used to clean up the previous
install before installing this version (and then updated to match
this new executables location).

Remove the path prefix on the .desktop file Exec key. This seens to
work - opening an .md by double-clicking in Nautilus still works,
so it must be finding the executable using the PATH. This means
the Exec field won't need to change when we move the executable scripts
like this commit does.