Remove the watch-markdown-gui wrapper script
This had been created earlier as an attempt to fix the problem
of leaving behind orphaned child processes after we exit, but
it failed to fix the problem, but got committed anyway intertwined
with the actual fix of a different way to kill child processes
in our bash scrip 'trap' on exit.

This commit just removes the needless wrapper script entirely.