Write output file to /tmp
This avoids cluttering the user's directory with unasked-for
HTML versions of their markdown files.

Instead we put the output HTML file into /tmp/watch-markdown/
with a filename that is visibly related to the full pathname of
the given input file (but with slashes replaced by backslashes,
to make a valid filename).

On exit, we delete the output html file.