Exit if our output file already exists
We interpret this to mean another watch-markdown process
is already running on our given input, so anything we might
do is redundant and possibly even harmful.

One problem is that the resulting error is not visible when
invoked from the desktop GUI. I wonder if we could system
notify it?