Silence errors from tidying our output file
If we are interrupted with Ctrl-C, then something seems to clean up our
output file for us. This surprises me, to be honest, but when we come to
'rm' the file in our EXIT trap, it does not exist.

Use 'rm -f' so that we no longer generate an error message in this case.