Display an error using system notifications
I was uncomfortable with the scenario in which we would refuse to run if
our output file already existed, under the assumption this meant that
another instance of watch-markdown was already running on the same input
file. When run from the command-line, this displayed an explanitory
error message, which was fine, but when run in the expected usual
manner, from the desktop GUI, the error to stderr was not displayed. For
us to not run with no feedback to the user seemed wrong.

So, now we display the error to stderr if we have a terminal attached,
and otherwise, send it to as a system popup notification.

This requires adding inotify-tools to our installed dependencies.