Switch from using 'inotifywait' to 'entr'
Ostensibly this is simpler, because we don't need to do the dance
of watching the file's parent directory and checking file names to
be able to follow the file when it is overwritten by a different
file, in the very common case of a Vim save or similar.

But, it's had a bunch of knock-on effects. The previous method of
killing our children used in cleanup didn't work, perhaps because we now
have extra processes that are not directly our jobs. (but the fix I have
put in place here does not attempt to recursively reap our tree of
children - it just gets the 1st generation direct children.)

And I still haven't gotten the quoting working for filenames with spaces
in. It's possible this never worked. Consider going back to the old
implementation to test it, and if that works, or can easily be made to
work, then just use inotifywait instead.